+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add settings support for multi-backend Logtalk executable arguments
 * Add quick fixes for some linter errors and warnings
 * Add quick fixes for some tests reporter warnings
+* Add quick fixes for some dead code scanner warnings
 * Add code navigation support for `(@)/1` goals
 * Update "Go to Symbol in Workspace..." support to also include predicate clauses and non-terminal rules
 * Improve performance of "Go to Symbol in Workspace..." and "Go to Symbol in Editor..." support